Adds distortion correction to Leap images
This change adds the distortion correction image to the images produced
by the LeapDevice. This is used to correct the lens distortion in the
Leap infrared images. In addition, a utility function is added to do the
distortion correction. This is distortion correction is purposly not
done in the device since it is an expensive calculation, and should only
be done when needed, in the simulation side.