Fixes images from LeapDevice in Debug
In debug mode, the images from the LeapDevice were not being provided,
this also happened rarely in Release mode. It turns out there was a race
condition where sometimes POLICY_IMAGES was being set to early. This
change sets the POLICY_IMAGES when the leap controller connects
(onConnect), resolving the issue.