SparseMatrix: fix mismatch between searchLowerIndex's return and Index.
searchLowerIndex returns type Eigen::Index, which seems to be ptrdiff_t.
Index (the third tparam for SparseMatrix) defaults to int.
The assignment of innerFirstElement = searchLowerIndex() thus throws
a conversion warning.  Since we know the type that the SparseMatrix
is using for its indices, we can static_cast to that type.