CollisionRep::update only calls getTransformed if the shape changed.
This means that in the UpdateCollisionRep Computation, only the
transformable shapes that have changed will incur the computational
expense of getTransformed.

Shape has getVersion, which can signal if the shape has changed.

Vertices has getUpdateCount, which counts the number of calls to update,
except for the update inside getTransformed.  This way, MeshShape's
getVersion can return getUpdateCount, but a transformed MeshShape
will have the same version as the instance that created it.

Similarly, CompoundShape tracks its version number, incrementing it
each time the shape changes, except in getTransformed.

Future work: look at DeformableCollisionRep too.