Fix bug, RecordPose was erasing the file in its initialization, causing trouble.
The issue was the following:
If using a yaml file to list all the devices to use and a ReplayPoseDevice is
listed after a DeviceFilter using a RecordPose with the same filename as the
ReplayPoseDevice, by the time we get to the ReplayPoseDevice, the file pointed
to by the RecordPose was emptied.

Solution: RecordPose has to open the file for writing (and emptying it) only
after the initialize. It could be in the main loop (filterInput method) or once
a InputConsumer gets connected (initializeInput). initializeInput happens only
once, so we chose this solution.