Updates NovintScaffold for the new 7Dofs.
The new 7Dof Falcons are a redesign. They must use
the osg_64.dll library to read in the roll angle and the
toolDof (i.e., grasper angle).  They have more torque.
And they apparently have different angular offsets.

Note that the new 7dof Falcons need to be accessed via
Simquest::NovintOsgDevice (which brings in the osg_64.dll).

One of the complications is that
the OSG Falcons are currently reporting angular ranges for pitch and yaw
that are incorrect.  These values look right when the tools are at the
center of the workspace, aligned parallel to the graound, and pointed into
the screen.

Flip some torques on left-handed 7dof Falcons...otherwise they're the
wrong direction.

I'm disabling the roll torque for now because the roll sensor is jittery.
To re-enable it, we'll have to add some smoothing or something.

Also, removes unnecessary linebreak in Novint log message.
This message can get displayed multiple times at startup.  The linebreak
makes scanning through the messages more difficult.