Settable limit for the applyCorrection step on Deformables.
We often are able to "blow up" our FemRepresentations if we deform them
too much, or put them in a situation where the MLCP can't find a nearby
solution.  If they "blow up" enough they'll go invalid, but oftentimes
that's too late.  Here's we can set a limit for how much each DOF can
change in a single applyCorrection, and if the correction is any bigger
we just discard it.

Note that this has no direct effect on FreeMotion, so over time the
deformable will relax into its rest state.