The MLCP results are discarded unless it returns a valid convergence.
The contact tolerance is now used in the determination of valid
convergence.

This change may naturally result in a loosening (i.e., increase) of the
contact tolerance to avoid "pretty good" solutions from being invalid.
So a second contact tolerance has been added.  More iterations will be
used to try and reach the tighter tolerance (but without the tighter
tolerance affecting the solution validity).