The VTC estimates velocity if it is missing.
The VTC's force calculations rely on the linear velocity and
angular velocity.  If these values are missing, the
proportional-derivative-control scheme goes unstable for damping gains
that should be reasonable.  Therefore, we either have to
use just a proportional-controller (ignoring velocities and damping) or
estimate the input velocity.

Estimating the input velocity will enable a higher performance system
response.  Unfortunately this simple estimate will generate high
frequency noise, so it is recommended that the input device provide
velocity measures if possible, or a more optimal filter (e.g., Kalman)
is used to add a velocity estimate to the input DataGroup before that
data is passed to the VTC.

Note that if we don't get either of the velocities, we reset both
Jacobian entries.  Using the Jacobians in the device scaffolds to
improve haptic feel requires both positions and velocities, otherwise
the response will be unbalanced (e.g., not enough damping at the
PD controller in the device loop, even if there's enough in the VTC).