Fix bug in Math::Geometry barycentricCoordinates, normal was not an unit vector.
The doxygen documentation says that the method called expects an unit vector,
but we are passing a non-unit vector.
In practice, it is not a bug, as it still calculates the correct barycentric
coordinates.

The problem comes in for small triangles, the normal calculated is a very small
vector, which scales down the signedAreaX2 variable in the method
barycentricCoordinates taking a triangle normal. This result in a comparison
with an epsilon that can fail more easily (less robust) and return nan for
triangles that are not degenerate, but simply small.