Add isApprox to DataStructures::Location
This is useful for filtering duplicate contacts in ccd mode.
MeshShape ccd collisions occurring at nodes or edges will be duplicated between
neighbors primitives (triangles) and this method will be useful to detect
duplicate contact to be filtered. In ccd, the collision will be find at the time
of impact, so neighbors triangle may find the exact same collision information.