Allow adding of Components in a SceneElement constructor
Defer setting of the SceneElement on a component until initialization of
the component. This allows a SceneElement to add components to itself in
it's constructor. You couldn't do this before, because the shared_ptr of
the scene element isn't available until after construction. So
SceneElement::getSharedPtr is invalid in the constructor (which
addComponent used to need).

SceneElements like Blocks::SphereElement, add their components in
doInitialize because of this and you can't access the elements
components until after it is added to the scene and initialized, which
isn't ideal. ie, when components are added in doInitialize you can't do
this:

    auto sphere = std::make_shared<SphereElement>("sphere");
    sphere->getComponent("Physics")->setValue("Density", 10.0);