CompoundShapeContactCalculationTests tests some contacts.
Previously it used zero-volume shapes that did not generate contacts.

The expected contacts must have their penetration points adjusted
by the pose of the subshape inside the CompoundShape, because the
CompoundShape's penetration points are with respect to the
CompoundShape, not the subshapes.

The tests are also changed such that the contact normals are
not aligned with any x/y/z axis to make the test slightly more
thorough.