Fix collision not being cleared on inactive representation
If a Collision::Representation is set inactive, it's current collisions
will not be cleared. So from then on, the inactive collision
representation will report the contacts it had when it was set inactive.
This fix adds a unit test that demonstrates this bug, along with the
fix. To clear/publish all collisions in PostUpdate, regardless if the
representation is active or not.