Fix failing SinkTest.Update
The sink test update was failing occasionally, due the collision of the
particle and sink (mesh) not being detected. This was likely due to the
particle collision radius being too small, and the collision being
missed. This change increase the size of the particle, and ensures it
collides on the first runtime step.

Also, be a bit safer by ensuring the runtime does two steps, so the
particle can be removed.