Remove nodeMeshLocalCoordinate from Location
nodeMeshLocalCoordinate and index both provide the same information in
DataStructures::Location, a 1D IndexedLocalCoordinate. Some
representations (MassSpringRepresentation) were looking for both of them
for node locations, and the others were only looking at the
nodeMeshLocalCoordinate. This change just removes the
nodeMeshLocalCoordinate since the index is much easier to use. ie, this:

    DataStructures::Location location(nodeIndex);

instead of this:

    DataStructures::Location location(
        DataStructures::IndexedLocalCoordinate(nodeIndex, Math::Vector()),
        DataStructures::Location::NODE);