Add plasticity to Fem2DElementTriangle.
This requires splitting the strain in local coordinates, between the membrane
part and the plate part. The final local force is rotated back into global
coordinate afterward.
This exposes much more terms in the force/stiffness calculation, which makes the
overall calculation a bit heavier, but there is no way around it. The plastic
calculation needs to evaluate the strain, which is local and hidden deep in
the equation.