Adds ImageMap class
This adds a new ImageMap class that is similar to Eigen's Map class. It
will wrap a raw data pointer (not copy) so it can act like our Image
class.  Any changes made to the ImageMap class will be reflected in the
original data.

Since the Image and ImageMap have a similar interface, an ImageBase
class is also created.