Reorder processComponents and processBehaviors
In the component managers, run processBehaviors before
processComponents.  Often times behaviors add components to the scene,
and it is desirable for these these components are available on the
current manager step, instead of having to wait until the next step.

In addition, behaviors now don't interact with components just after
they are added (in an initial state) and before the managers get a
chance to update them.