Change Mlcp friction coefficient array indexing to atomicConstraintId
For frictional constraint, the very first entry will store the friction
coefficient for the entire constraint. The other entries will remain unused.
This is useful to define the friction coefficient in the various
Constraint::doBuild method without having to add a new parameter to know which
constraintId it is. We can use directly the atomicConstraint index.