Simplify InputComponent and OutputComponent
In an attempt to make the InputComponent and OutputComponent easier to
inherit from, merge the InputConsumer and OutputConsumer into the
InputComponent and OutputConsumer, respectively. This seperation was
making deriving from the OutputComponent difficult. When subclassing
the OutputComponent, it is usually necessary to create a new
OutpuConsumerInterface derived class. This resulted in a bunch of
undesirable static casts in the OutputComponent.

Also, remove isDeviceConnected to make this transition easier. This
function was not being used anywhere, currently.