RenderTestCcdSuture: Fix the segment tests to work with the CCD
The Fem 1D self intersection and Ccd render tests now reflect the
changes necessary to work with the CCD code. Note that there is a
relationship between the radius of the colliding 1D segment and the
frame rate. The thinner the segment, the higher the physics loop
rate required for successfully maintaining the collisions.