CombiningOutputComponent combines multiple OutputComponents into one.
Devices can only have one OutputProducerInterface, which is generally
provided via an OutputComponent, and which in turn has its data set via
a Behavior like the VirtualToolCoupler.

In order to drive a device with multiple outputs, e.g., two or more
VirtualToolCouplers, then the CombiningOutputComponent can be hooked
up to the device, and the sub-OutputComponents can be set on it.
The default functor that handles the data combining will copy the data
from the first OutputComponent, reset any spring or damper Jacobians
(which could be combined but we have not done that work), and then
add any forces or torques from the subsequent OutputComponents.

User-defined functors could perform alternative data combinations.