Fixes incorrect unit test, CombiningOutputComponentTest.Serialization.
The element provides components in a different order than they are
stored in the CombiningOutputComponent, so the unit test should not
expect the vectors to be the same (the values may be in a different
order).

Also, use an unordered_set it CombiningOutputComponent::setOutputs
because we don't need the ordering.