Fixes incorrect unit test, CombiningOutputComponentTest.Serialization.
This bug in the unit test has the same cause as the previous commit:
The element provides components in a different order than they are
stored in the CombiningOutputComponent, so the unit test should not
expect the vectors to be the same (the values may be in a different
order).

Since it's the first OutputComponent's data that gets copied in the
default functor, we need to make sure we know which OutputComponent
is first in the CombiningOutputComponent.