Add Fixed/Rigid/Fem constraint fixed rotation vector.
In this implementation, the fem is the only active representation, the
rigid/fixed ones are only controlling the rotational dof and not receiving any
force feedback from this constraint.

This allows the user to control the orientation of a Fem beam/triangle from a
fixed representation or a rigid representation (controlled by a VTC or not).