CombiningOutputComponent maintains its vector<weak_ptr<OutputComponent>>
If a weak pointer is no longer lockable, it gets removed from the vector
held by the CombingingOutputComponent.  With this change, the functor
gets shared_ptrs instead of weak_ptrs.

The default combining functor is also simplified.  Now it only
accumulates the force and torque values.  This way, if the first
OutputComponent goes away, the CombiningOutputComponent will still
provide data.  (Previously the default functor needed the first
entry in the output vector to provide data.)