Removes OutputComponent::getData, which duplicates requestOutput.
When OutputComponent had-a OutputProducerInterface, we needed a function
that could directly access the producer's data so that other Components
could use that data.  See CombiningOutputProducer.

Now that OutputComponent is-a OutputProducerInterface, Components can
directly call requestOutput on it, so getData is no longer necessary.

Also, getData had a bug because before setData is called the DataGroup
will be empty, and so getData would return an empty DataGroup.
If return value optimization was not being used, then this code:
DataGroup data = outputComponent.getData();
would cause an assertion due to the copy construction based off an
empty (invalid) DataGroup.