InputManager only holds onto Input/Output Components that correctly add.
If addInputComponent/addOutputComponent fails, the InputManager does
not hold onto the component in m_inputs/m_outputs.

Also, OutputComponents can use a device name of the empty string, which
adds them to the InputManager without connecting them to a device.
OutputComponents combined with a CombiningOutputComponent should do so;
i.e., just don't set a device name on them.  As a consequence, we cannot
add an actual device with an empty name.