Fix bug in rotation vector (the constraint violation must be a rotation vector)
So far, the rotation vector constraint was calculating the violation as a simple
difference of rotation vector. While it should be the rotation vector needed
to go from one rotation to another one. This ensures a proper violation but also
handle the specificity of the rotational space (rotation vector can flip past a
norm of PI).