CcdCollisionLoop: Make CCD available in the same manner as DCD
By not restoring the CCD collisions in CcdCollisionLoop, and by not
providing the collision position, we were making CCD less inter-
changeable with DCD. This corrects those problems.

As an incidental change, it is often useful to see the actual
form of a segment mesh intead of an interpolated form. The
changes to CardinalSpline allow one to visualize uninterpolated
segment meshes easily by specifying 0 subdivisions.