Prevent getData asserting with unconnected InputComponents
If getData is called on an unconnected InputComponet, an assert occurs
while trying to copy uninitiliazed NamedData. Not a very helpful runtime
error. This change prevents the copy, and assert, if something attempts
to getData on an unconnected InputComponent. getData will just return
without changing the passed in DataGroup. The nice thing about this
change, is elements with InputComponents can be added to the scene and
not have a device hooked up.