7dof Novints use the same device workspace as other devices.
All of our devices should have +z towards the user, +y up, and +x to
the user's right.  That matches what we have with the MultiAxisDevice
the PhantomDevice, and the 3dof NovintDevice.  We expect to use the
7dof OSG Falcons with the left Falcon to the user's left, and the
right Falcon to the user's right, so they should be setup to use
the same workspace to simplify switching between devices.

Also, the home pose for the 7dof Falcons should be with the handle
of the tool parallel to the ground and pointed towards the user, so
that the tip (business end) of the tool points along the -z axis.

Also improves NovintScaffold logs when missing configuration files.

Also since the OSG Falcons do not provide their roll angle via the HDAL
(instead they use an Arduino interface that we connect to in
NovintOsgDevice), the NovintScaffold does not use the change in
orientation to calculate force/torque via Jacobians.