setMaterial on Graphics Representations to take Components
Change the interface of Graphics::Representation to take
Framework::Components instead of Graphics::Material. This makes it
conveniant to set a material (which is a Component) on a representation
when it is looked up from the Scene. This will probably have to be done
in the future when Materials are serialized anyways.