Adds addUniform overload to OsgMaterial and OsgRepresentation
Adds a new form of addUniform to both an OsgMaterial and
OsgRepresentation. This allows you to create and set the value of the
uniform in one call. ie:

  material->addUniform("float", "shininess", 0.9f);

 instead of:

  material->addUniform("float', "shininess");
  material->setValue("shininess", 0.9f);

Switches addUniform to ASSERT when the uniform type is not OsgUniform,
or the glsl type is not available. This was the original behavior of
OsgRepresentation, and is now being uniformily used in OsgMaterial as
well.