NovintScaffold does not know velocity & should not use damper Jacobian.
This change should not affect any simulations because the NovintScaffold
does not provide velocities and therefore should not be receiving
INPUT_LINEAR_VELOCITY or INPUT_ANGULAR_VELOCITY in the output datagroup.
However, if it did get those data entries, it would calculate a
force/torque as though the velocity went to zero in this callback.

We may want to add a velocity filter to NovintDevice/Scaffold.
(See VelocityFromPose DeviceFilter.)  Theoretically we would see
improved haptic feel, but the filter might require tuning for
optimal performance (e.g., minimizing lag vs maximizing accuracy).