Remove exception from StreamOutput destructor
Throwing an exception from a destructor is a bad idea, and actually in
C++11, destructors are marked as noexcept. GCC6 correctly identitified
this occuring in the StreamOutput destructor and produced a warning. To
fix this, I just removed the destructors from StreamOutput and
FileOutput as they don't do anything useful and to follow the "Rule of
Five".