Fix precision issues in VectorTests
Due to GCC6, or maybe the latest Eigen (3.3~beta2), the VectorTests
MultiplyVectorScalar and MultiplyScalarVector are failing with float
vectors. The difference between the actual and expected result was about
2e-6, larger than the epsilon of 1e-6. This is actually, not unexpected
as single precision float have 6-9 significant digits. With these
particular test, 6 floats are multipled and then added, so floating
point imprecision is likely greater than 1e-6. To fix the issue, the
epsilon for these tests, and other tests involving sums, is increased to
5e-6.