CcdCollisionLoop: Fix different behavior when there are no CCD contacts
CcdCollisionLoop processes all contacts (CCD and DCD) if there is
any CCD collision. It did not process anything if there is no CCD
collision. This fix ensures that in the case that there are
no CCD contacts, at least one pass through the system is taken
to generate a DCD contact solution.