OdeState: add boundary conditions for statically defined dof
Degrees of freedom (dof) are defined as dynamic parameter of the physical model,
their values are computed with a ode solver.
But some models have statically defined dof, which are not listed as dynamic
parameters and are not determined by an ode solver, but on the contrary, defined
statically prior to the numerical integration. (example: instant propagation of
the twist along a 1d model).