Add get/setValue to SceneElement
This adds convenience functions to SceneElements to access properties on
its Components. All the forms of the Components get/setValue functions
are recreated on the SceneElement, but with an added parameter for the
component name. The functions also include checking if the component
actually exists. So instead of this:

  auto value = element->getComponent("Component")->getValue("Property");
  element->getComponent("Component")->setValue("Property", value);

You can more safely and conveniently do this:

  auto value = element->getValue("Component", "Property");
  element->setValue("Component", "Property", value);