ctest's memorycheck is less verbose.
By default, ctest sets valgrind to output all "lost" memory.  This
commit tells valgrind to only output the "definitely lost" and
"possibly lost" memory, not the "indirectly lost" memory.
Indirectly lost memory will be fixed when definitely lost memory is
fixed, so we don't need to see both, and the indirectly lost memory
is too voluminous.

This commit also removes the memorycheck suppressions file.  It does
not seem to be necessary at this time.