Revert "Make the groups inside of sceneelement a shared_ptr"
This reverts commit 3cb608b9e26c3adcaa1a7a7dbb2d1fd1fc7943b5. The
shared_ptrs in groups back to the SceneElement was causing a circular
reference. This is handled when SceneElements are in a Scene, becuase
the Scene will break the circular reference on its desctruction. But in
unit tests, SceneElements often are not added to a Scene, resulting in a
massive amount of leaked memory.

This change reverts the "groups" help by a SceneElement back to a vector
of strings, as it was previously.