Retire Representations with contacts in Unit Tests
Collision Representations that are in contact with other representations
hold shared_ptrs to each other. This is handled by calling "retire" on
representations before they are destructed. While this is handled by
managers properly in a simulation, a couple of Unit Tests don't use
managers, and where not properly retiring the representations. As a
result, memory was leaked.