Fix massive memory leak in OsgTextureCubeMap
The OsgTextureCubeMap was leaking hundreds of megabytes in the graphics
unit tests. This was because we allocated a buffer for an osg::Image,
but told the osg::Image not to handle deleting the memory
(osg::Image::NO_DELETE) while not deleting it ourselves. This change
correctly tells the created osg::Image to manage the resource
(osg::Image::USE_NEW_DELETE).

In addition, we were unnecessarily allocating space for the data twice.