Fix using uninitialized data in Fem3DElementCubeTests
Valgrind reported uninitialized data was being used in the
Fem3DElementCubeTests. This was because the test was trying to get the
volume of a Cube Element. These elements need to be initialized to setup
a couple of member variables (m_shapeFunctionsEpsilonSign,
m_shapeFunctionsEtaSign, m_shapeFunctionsMuSign) before calling
getVolume. The tests in question ensure that the element volume
calculation asserts when given bad element geometry. The fix is to just
call initialize instead, since "initialize" correctly sets up the member
variables, and also calculates the volume.