Fix Unit Tests failures uncovered by running through Valgrind
A couple of unit tests started failing when they were run through
Valgrind. Likely due to the slowness of running a test the memory
checker, that highlighted issues not seen before.

The fixes include:

* instead of sleeping some ammount of time in the
DcdCollisionTest.Deactive, just call step twice to ensure the managers
were run once.

* Added needed sleeps after stopping a thread, and before it is
desctructed in BasicThreadTests

* In the BoolToScalarTest, the slowness of running through valgrind
caused the update (and timer) to run significantly longer than the
sleep. As a result, just ensure the value change is larger or equal to
what you expect.

* Lastly, when launching the BasicThread, the thread was exiting because
it checks if m_isRunning is true in its main while loop. Since
m_isRunning is set to true right after the thread launch, this is
usually fine, but when running under Valgrind the thread ran all the way
through and exited before the line immediatly after the launch (which
set m_isRunning to true) was executed. The fix is to just set
m_isRunning to true immedialty before launching.