Fix failing Ccd unit tests due to floating point precision issues
GeometryTest.CcdIntersectionsPointTriangle unit test was failing because
of floating point precision issues in the CubicSolver. This increases
the epsilon used for a couple of isNearZero in the solver.

SegmentSegmentCcdStaticContactTests.SegmentSegmentCollisions was failing
with parallel segments because on a recent i386 debian unstable machine,
0.0 was evaluating to be >= 1e-18 in
SegmentSegmentCcdStaticContact.cpp:85. This checks if the segments are
parallel. 1e-18 is probably at or beyond double precision limit. To fix
this, the precision is changed to 1e-15.