Add new search path for GoogleMock
Google Mock/Test has changed the directory structure in the latest
release, 1.8. As a result the debian package places things in a slightly
different location. To automatically find GoogleMock installed from the
repository, the new location is added to FindGoogleMock.cmake. The old
location is left so 1.7 will still be found.