ContactCalculationTests can log the contacts.
It is hard to create poses and motions that result is the desired
contacts for testing, and then it is hard to determine the
global positions and normals of the contacts due to those
poses & motions.  This log makes that effort easier.

Note, normally log messages are not printed because TestingMain.cpp
sets the default output to NullOutput.