Adjust CubicSolver/Polynomial epsilons.
In trying to add a second edge-edge test case for
TriangleMeshTriangleMesh CCD, in which one of the two meshes
was stationary, the CubicSolver was providing erroneous roots.
Numerical precision is the problem, caused by a in ax^2+bx+c=0 being
small.

Setting the epsilon for the CubicSolver to that used by Polynomial
and PolynomialRoots fixes that problem, but causes poor
collision response in the Fem1D CCD render tests.  Tightening the
epsilons to 1e-15 makes those render tests look better again.