findRootsInRange01 uses larger epsilon for degenerate cases.
If the cubic is degenerate, then we use the same epsilon to
determine if the quadratic is degenerate.  This change fixes
a failure on 32 bit Debian in the TriangleMeshTriangleMeshCcd tests.
It does not appear to affect the Fem1D collision behavior.

Adds a unit test for TriMeshTriMeshCcd when rep#1 is stationary.
I wanted to add this earlier but was getting errors before the
change here for epsilon.

This commit also cleans up the logging for collision unit tests,
and shortens the duration of the new tri-tri ccd render tests.