Corotational FemReps calculate the compliance matrix asynchronously.
This change means that the compliance matrix is always stale by
one update.
Since the solve in the FreeMotion update takes about as long as the
corotational matrix multiplication, we can perform them in parallel
in half the time as in series.

We could launch the asynchronous task at a different time, such as
after PushResults, but it would still be stale to some degree,
and it is not obviously better to have it be stale only by the
FreeMotion step versus by a full update.

A significant amount of time is being spent copying the compliance
matrix out of the result, in this line:
m_complianceWarpingMatrix = m_complianceWarpingMatrixForCalculation;
Perhaps that can be improved.