Cmake setting for Runtime's threadpool size.
With hyperthreading enabled, the OS treats each physical core as
two (or more) logical cores.  However, those logical cores are
sharing the physical core's cache, which can cause performance to
decrease if the threads on the logical cores are fighting over
the cache.  Our ThreadPool defaults to using the number of
logical cores, but I am getting faster physics update rates
if I only use half that number (i.e., the number of physical cores
in my computer).

This makes the Runtime's ThreadPool size a cmake parameter for
easy editing.