Updates to Eigen 3.3.2
Eigen 3.3 deprecates the Eigen::SparseMatrix::Index, because there
are separate index types for general access and for access of the
data storage in a sparse matrix, respectively Eigen::Index
and the StorageIndex template parameter of SparseMatrix.

Furthermore, this commit uses Eigen::Ref for the dense submatrix
for Operation::add and Operation::assign.  That ref avoids a resize
and thereby provides a significant speed improvement.