Fem1DElementBeam constructs mass matrix as symmetric.
The matrix is symmetric, this just changes the assignments to
make that clear.

Also, adds Latex tags in OdeEquation::applyCompliance dox.

Ultimately my confusion was because OdeSolverLinearEulerImplicit
calculates the compliance matrix once, and then uses it to solve
for the change in velocities given the force...
via C.delta_x = force?
This calculation is an optimization because it means we don't need
to recalculate anything but the force.