CompoundShapeToGraphics has local offsets for the graphics.
This commit helps when the CompoundShape's subshape is posable while
the graphics rep's shape is not, or vice versa.

When CompoundShapeToGraphics gets the CompoundShape's subshape's pose,
it needs the pose of the subshape with respect to the CompoundShape.
That relative pose is then transformed by the
"offset" pose of the graphics rep to get the pose of the
graphics rep with respect to the scene element (i.e., its local pose).

If both the subshape and the graphics rep are meshes, and they have
the same origin, then no offset is necessary.  Ditto if both are
non-posable.

Without this change, a CompoundShape with a non-posable subshape
(e.g., sphere) would pose a mesh-based graphics rep so its origin was
at the center of the sphere.  Trying to drive a graphics rep of a
tool from a sphere shape collision rep would require re-modeling the
mesh to position its origin based on the local pose of the sphere.
With this commit, CompoundShapeToGraphics stores the initial
local pose of the graphics rep and uses it to offset the subshape and
graphics rep. Thus, by setting the graphics rep's initial local pose
as the inverse of the local pose for the subshape, the graphics
can be "placed" at any pose relative to the subshape.