VisualizeContactsBehavior correctly orients vectors.
Previously the element's pose was used to position the contact point,
but not to orient the direction of the vector. So an element with a
rotation would have vectors in the wrong directions.

Colors the two vectors differently, which can help differentiating
the contacts on the two representations.

Also, adds VisualizeContactsBehavior to the RenderTestRigidBodies,
which serves as a good render test for the behavior.
VisualTestFallingSphereOnMesh in particular shows the bug fixed by
the changes to VisualizeContactsBehavior.
Spreads out the spheres in the VisualTestFallingSphereOnMesh, so there
are more collisions.