Fixes the expected penetration depth of the new test.
This commit does not fix the bug that causes assertions.
Instead, it makes the test provide the correct expected penetration
depth for the variations that do not assert.

However, the triangle point is not equal to the expected triangle
point.  To make them equal, use something like
Vector3d(1e-12, 1e-12, 0.0) for the expected triangle point.