Potential fix for the triangle-capsule assert.
clipSegmentAgainstTriangle does not need epsilons.  If the segment
passes through the plane, then it should be clipped,
otherwise it shouldn't.

Also, fixes the penetration point on the triangle when a capsule
endpoint is on the triangle or the initial triangle penetration point
is on a triangle edge.

And fixes the expected penetration points for the new test case.