Adds distanceLinePlane to Geometry, and improves point-triangle calls.
isPointInTriangle now accounts for triangle size for precision purposes.
It uses barycentric coordinates to determine if a point is inside of
a triangle, but barycentric coordinates are unitless
(they range from 0 through 1 inside the triangle) and mean more or less
based on the size of the triangle.  Scaling the coordinates by a
distance turns them into a distance so they can be compared with
the distance epsilon. The distance used is the length from the vertex
to the midpoint on the opposite side.

Similarly, isPointOnTriangleEdge uses distancePointSegment so that the
comparisons with epsilon are distances.