axisTouchingTriangle in TriangleCapsuleContact finds better contacts.
This commit addresses the recent additions to the unit tests.

Previously axisTouchingTriangle used the closest point on the segment
to the triangle, and took the intersection on the capsule in the
negative direction of the triangle normal.

Now, it checks if the penetration point on the capsule is on the
hemisphere about the bottom endpoint. If not, it checks if the
penetration point is on the cylinder.  Finally, it checks if the
penetration point is on the sphere about the top endpoint.

This code draws on the code in axisThroughTriangle, but they have
different assumptions. It may be possible to combine them.

This commit also adds more info to asserts messages.