DriveElementFromInputBehavior is a physics behavior.
Endovascular was generating sporadic assertions in the
contact calculations because DriveElementFromInputBehavior
was running in the BehaviorManager.

It appears that DriveElementFromInputBehavior was setting
the SceneElement's pose after UpdateCollisionData had updated
the mesh vertices, but before UpdateDcdData had updated
the mesh normals. Then in the next physics loop the vertices
would get updated, but not the normals.