TriangleCapsuleContactCalculation handles coplanar segment.
If the segment is co-planar with the triangle, and the endpoint
initially labeled "bottom" is outside of the triangle, then the
contact calculation needs to check if the "top" is inside the
triangle or on the triangle edge.  Only if that is false should
the next step be axisTouchingTriangleWithBottomOutside().

axisTouchingTriangleWithBottomOutside assumes that the "bottom" endpoint
is deeper than the "top" endpoint, so it may not correctly
handle co-planar situations.