SlidingConstraint uses the previous suture localization for friction.
Previously, the constraint first moved the localization on the suture
to be close to the hole in the tissue, and then it used that new
localization to determine friction forces.  That approach ignores
any free motion changes that move the suture along the sliding
direction.

This approach is still imperfect because no friction forces are
generated for motions parallel to the sliding direction, so if the
suture starts sliding through at an angle, only some of that motion
generates friction.